feat: add confirmation dialog for control scheme deletion

+++ src/app/control-schemes/control-scheme-list/control-scheme-list.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, OnDestroy, TemplateRef, ViewChild } from '@angular/core';
 import { LetDirective, PushPipe } from '@ngrx/component';
 import { NgForOf, NgIf } from '@angular/common';
-import { TranslocoModule } from '@ngneat/transloco';
+import { TranslocoModule, TranslocoService } from '@ngneat/transloco';
 import { Store } from '@ngrx/store';
 import { MatButtonModule } from '@angular/material/button';
 import { RouterLink } from '@angular/router';
@@ -13,6 +13,7 @@ import { FeatureToolbarService } from '@app/shared';
 import { RoutesBuilderService } from '../../routing';
 import { ControlSchemeListItemComponent } from '../control-scheme-list-item';
 import { CONTROL_SCHEME_ACTIONS, CONTROL_SCHEME_CONFIGURATION_STATE_SELECTORS, CONTROL_SCHEME_SELECTORS, ControlScheme } from '../../store';
+import { ConfirmDialogService } from '../../shared/confirm-dialog';
 
 @Component({
     standalone: true,
@@ -42,7 +43,9 @@ export class ControlSchemeListComponent implements OnDestroy {
     constructor(
         private readonly store: Store,
         private readonly featureToolbarService: FeatureToolbarService,
-        protected readonly routesBuilderService: RoutesBuilderService
+        protected readonly routesBuilderService: RoutesBuilderService,
+        private readonly confirmDialogService: ConfirmDialogService,
+        private readonly translocoService: TranslocoService
     ) {
     }
 
@@ -57,13 +60,29 @@ export class ControlSchemeListComponent implements OnDestroy {
 
     public ngOnDestroy(): void {
         this.featureToolbarService.clearConfig();
+        this.confirmDialogService.hide(this);
     }
 
     public trackSchemeById(index: number, scheme: ControlScheme): string {
         return scheme.id;
     }
 
-    public onDelete(id: string): void {
-        this.store.dispatch(CONTROL_SCHEME_ACTIONS.delete({ id }));
+    public onDelete(
+        id: string,
+        name: string
+    ): void {
+        this.confirmDialogService.show(
+            this.translocoService.selectTranslate('controlScheme.deleteSchemeConfirmationTitle', { name }),
+            this,
+            {
+                content$: this.translocoService.selectTranslate('controlScheme.deleteSchemeConfirmationContent'),
+                confirmTitle$: this.translocoService.selectTranslate('controlScheme.deleteSchemeConfirmationConfirmButtonTitle'),
+                cancelTitle$: this.translocoService.selectTranslate('controlScheme.deleteSchemeConfirmationCancelButtonTitle')
+            }
+        ).subscribe((isConfirmed) => {
+            if (isConfirmed) {
+                this.store.dispatch(CONTROL_SCHEME_ACTIONS.delete({ id }));
+            }
+        });
     }
 }
